add remove file from cloudinary

+++ src/Controller/PhotoHotelController.php
@@ -123,12 +123,12 @@ class PhotoHotelController extends AbstractController
             throw $this->createAccessDeniedException();
         }
         if ($this->isCsrfTokenValid('delete'.$photo->getId(), $request->request->get('_token'))) {
-            $folder = $this->getParameter('kernel.project_dir').'/public/uploads/photos/';
-            $path = $folder . $photo->getLien();
-            $filesystem = new Filesystem();
-            $filesystem->remove($path);
-
-            $photoRepository->remove($photo);
+            $path = $photo->getLien();
+            $result = (new ImageUploader())->remove(substr($path, 0,  strrpos($path, ".")) );
+            if($result['result']=='ok'){
+                $this->addFlash('success', 'Votre image a été supprimée'.' ('.substr($path, 0,  strrpos($path, ".")).')');
+                $photoRepository->remove($photo);
+            }
         }
 
         return $this->redirectToRoute('app_photo_index', [
